extracted TagAutocomplete functionality into its own file
