+++ packages/desktop-client/src/components/transactions/TransactionsTable.tsx
@@ -83,6 +83,7 @@ import { AccountAutocomplete } from '#components/autocomplete/AccountAutocomplet
 import { Autocomplete } from '#components/autocomplete/Autocomplete';
 import { CategoryAutocomplete } from '#components/autocomplete/CategoryAutocomplete';
 import { PayeeAutocomplete } from '#components/autocomplete/PayeeAutocomplete';
+import { TagAutocomplete } from '#components/autocomplete/TagAutocomplete';
 import { getStatusProps } from '#components/schedules/StatusBadge';
 import type { StatusTypes } from '#components/schedules/StatusBadge';
 import { DateSelect } from '#components/select/DateSelect';
@@ -407,12 +408,12 @@ function StatusCell({
           ':focus': {
             ...(isPreview
               ? {
-                  boxShadow: 'none',
-                }
+                boxShadow: 'none',
+              }
               : {
-                  border: '1px solid ' + theme.formInputBorderSelected,
-                  boxShadow: '0 1px 2px ' + theme.formInputBorderSelected,
-                }),
+                border: '1px solid ' + theme.formInputBorderSelected,
+                boxShadow: '0 1px 2px ' + theme.formInputBorderSelected,
+              }),
           },
           cursor: isClearedField ? 'pointer' : 'default',
           ...(isChild && { visibility: 'hidden' }),
@@ -555,8 +556,8 @@ function PayeeCell({
           ':hover': isPreview
             ? {}
             : {
-                border: '1px solid ' + theme.buttonNormalBorder,
-              },
+              border: '1px solid ' + theme.buttonNormalBorder,
+            },
         }}
         disabled={isPreview}
         onSelect={() =>
@@ -1462,11 +1463,11 @@ const Transaction = memo(function Transaction({
             value={
               matched
                 ? // TODO: this will require changes in table.tsx
-                  ((
-                    <SvgHyperlink2
-                      style={{ width: 13, height: 13, color: 'inherit' }}
-                    />
-                  ) as unknown as string)
+                ((
+                  <SvgHyperlink2
+                    style={{ width: 13, height: 13, color: 'inherit' }}
+                  />
+                ) as unknown as string)
                 : undefined
             }
           />
@@ -1733,11 +1734,11 @@ const Transaction = memo(function Transaction({
             valueStyle={
               !categoryId
                 ? {
-                    // uncategorized transaction
-                    fontStyle: 'italic',
-                    fontWeight: 300,
-                    color: theme.formInputTextHighlight,
-                  }
+                  // uncategorized transaction
+                  fontStyle: 'italic',
+                  fontWeight: 300,
+                  color: theme.formInputTextHighlight,
+                }
                 : valueStyle
             }
             onUpdate={async value => {
@@ -1972,66 +1973,6 @@ function NotesCell({
   onExpose,
 }: NotesCellProps) {
   const [inputValue, setInputValue] = useState(value);
-  const inputRef = useRef<HTMLInputElement | null>(null);
-  const [cursorPosition, setCursorPosition] = useState(0);
-  useEffect(() => {
-    inputRef.current?.setSelectionRange(cursorPosition, cursorPosition);
-  }, [cursorPosition, setCursorPosition]);
-
-  useEffect(() => setInputValue(value), [value, setInputValue]);
-  const tagQuery = useTags();
-  const tagOptions =
-    tagQuery.data?.map(tag => ({ id: tag.tag, name: '#' + tag.tag, tag })) ??
-    [];
-
-  function onSelect(
-    optionId: string,
-    value: string,
-    e?: KeyboardEvent<HTMLInputElement>,
-  ) {
-    const [start, end] = getCurrentWordRange(value, cursorPosition);
-    const option = tagOptions.find(o => o.id === optionId);
-    if (option) {
-      const newValue =
-        value.slice(0, start) + option.name + ' ' + value.slice(end);
-      setInputValue(newValue);
-      setTimeout(() => setCursorPosition(start + option.name.length + 1));
-
-      // only stop event propagation (i.e. table navigation) when we want to do
-      // autocomplete things. If we don't choose an option, then we want to treat
-      // this as a regular input field and do table navigation.
-      e?.stopPropagation();
-    } else {
-      onUpdate(value);
-    }
-  }
-
-  function onKeyUp(e: KeyboardEvent<HTMLInputElement>) {
-    setCursorPosition(e.currentTarget.selectionStart ?? 0);
-  }
-
-  function onKeyDown(e: KeyboardEvent<HTMLInputElement>) {
-    const highlightedEntries = document.querySelectorAll(
-      'div[data-tag-highlighted=true]',
-    );
-
-    if (e.key === 'Escape') {
-      // reset to initial value
-      setInputValue(value);
-    }
-    if (e.key === 'Tab' && highlightedEntries.length) {
-      const selectedId =
-        highlightedEntries.item(0).getAttribute('data-tag-option-id') ?? '';
-      e.preventDefault();
-      e.stopPropagation();
-      onSelect(selectedId, inputValue, e);
-    } else if (e.key === 'Enter' || e.key === 'Tab') {
-      // set to current value. For some reason this is
-      // is not getting caught by the onBlur handler
-      // likely because of Autocomplete complexity
-      onUpdate(inputValue);
-    }
-  }
 
   return (
     <CustomCell
@@ -2046,101 +1987,16 @@ function NotesCell({
       onExpose={onExpose}
       onUpdate={onUpdate}
       onFocus={() => setInputValue(value)}
-      onBlur={() => onUpdate(inputValue)}
-      onKeyDownCapture={onKeyDown}
     >
       {({ inputStyle, onKeyDown, onBlur }) => (
-        <Autocomplete
-          strict={false}
+        <TagAutocomplete
           value={value}
-          embedded={false}
-          highlightFirst
-          clearOnBlur={false}
-          closeOnBlur
-          openOnFocus={false}
-          onSelect={onSelect}
-          inputProps={{
-            ref: inputRef,
-            onBlur,
-            onKeyUp,
-            onKeyDown,
-            style: inputStyle,
-            value: inputValue,
-            onChange: v => setInputValue(v.target.value),
-          }}
-          suggestions={tagOptions}
-          renderItems={(items, getItemProps, highlightedIndex) => {
-            return (
-              <View style={{ paddingLeft: 4, paddingRight: 4 }}>
-                <View
-                  style={{
-                    overflowY: 'auto',
-                    willChange: 'transform',
-                    padding: '5px 0',
-                    maxHeight: 175,
-                  }}
-                >
-                  {items.map((item, idx) => (
-                    <div
-                      key={item.id}
-                      {...getItemProps({ item })}
-                      onClick={() => onSelect(item.id, inputValue)}
-                      role="button"
-                      data-tag-option-id={item.id}
-                      data-tag-highlighted={
-                        highlightedIndex === idx ? 'true' : undefined
-                      }
-                      className={cx(
-                        css({
-                          backgroundColor:
-                            highlightedIndex === idx
-                              ? theme.menuAutoCompleteBackgroundHover
-                              : 'transparent',
-                          color:
-                            highlightedIndex === idx
-                              ? theme.menuAutoCompleteItemTextHover
-                              : theme.menuAutoCompleteText,
-                          padding: 4,
-                          borderRadius: 4,
-                          border: 'none',
-                          font: 'inherit',
-                        }),
-                      )}
-                    >
-                      <NotesTagFormatter notes={item.name} />
-                    </div>
-                  ))}
-                </View>
-              </View>
-            );
-          }}
-          filterSuggestions={(options, inputValue) => {
-            if (inputValue.trim() === '' || !cursorPosition) {
-              return [];
-            }
-            const currentWord = getCurrentWord(
-              inputValue,
-              cursorPosition,
-            )?.toLowerCase();
-            if (!currentWord || !options || currentWord.charAt(0) !== '#') {
-              return [];
-            }
-            const currentWordNoHash = currentWord.slice(1);
-            return options
-              .filter(o => o.id.toLowerCase().includes(currentWordNoHash))
-              .sort(({ id: a }, { id: b }) => {
-                const aStarts = a.toLowerCase().startsWith(currentWord);
-                const bStarts = b.toLowerCase().startsWith(currentWord);
-                if (aStarts && !bStarts) {
-                  return 1;
-                } else if (!aStarts && bStarts) {
-                  return -1;
-                }
-                const compare = a.toLowerCase().localeCompare(b.toLowerCase());
-                return compare > 0 ? 1 : compare < 0 ? -1 : 0;
-              })
-              .slice(0, 10);
-          }}
+          inputValue={inputValue}
+          setInputValue={setInputValue}
+          inputStyle={inputStyle}
+          onBlur={onBlur}
+          onKeyDown={onKeyDown}
+          onUpdate={onUpdate}
         />
       )}
     </CustomCell>
@@ -3174,10 +3030,10 @@ export const TransactionTable = forwardRef(
       fields = item?.is_child
         ? ['select', 'payee', 'notes', 'category', 'debit', 'credit']
         : fields.filter(
-            f =>
-              (props.showAccount || f !== 'account') &&
-              (props.showCategory || f !== 'category'),
-          );
+          f =>
+            (props.showAccount || f !== 'account') &&
+            (props.showCategory || f !== 'category'),
+        );
 
       if (item?.id && isPreviewId(item.id)) {
         fields = ['select'];
@@ -3489,9 +3345,9 @@ export const TransactionTable = forwardRef(
           t =>
             t.parent_id &&
             t.parent_id ===
-              (transaction?.is_parent
-                ? transaction?.id
-                : transaction?.parent_id),
+            (transaction?.is_parent
+              ? transaction?.id
+              : transaction?.parent_id),
         );
 
         const emptyTransactions = siblingTransactions.filter(
